create

+++ model/contacts/removeContactModel.js
@@ -1,11 +1,12 @@
 const Contact = require('../../schemas/contacts')
+const { NOT_FOUND } = require('../../helpers/constants')
 
-const removeContactModel = async (contactId) => {
+const removeContactModel = async (userId, contactId) => {
   try {
-    const contact = await Contact.findByIdAndDelete({ _id: contactId })
+    const contact = await Contact.findByIdAndDelete({ _id: contactId, owner: userId })
     return contact
   } catch {
-    throw new Error('404')
+    throw new Error(NOT_FOUND)
   }
 }
